Clarify that overriding policy document statements can be added to current document if `sid` does not match
In the description of `iam_policy_document` data source's
`override_json` argument, state specifically that any statements in the
override document that have `sid`s that do _not_ match a statement in
the current document will be added to the current document, instead of
ignored. Before this change, there was no description of how
non-matching statements in the `override_json` were used, so the
behavior was unclear: the opposite and incorrect conclusion that these
statements were ignored was equally consistent with the documentation.

This behavior is covered by an acceptance test, so from the coverage I'm
inferring that this behavior is both intended and supported.